updated time loop, this is gonna take a while

+++ routes/patients.js
@@ -2,6 +2,7 @@ const express = require('express');
 const Router = express.Router();
 const passport = require('passport');
 const Patient = require('../models/patients');
+const CalEvent = require('../models/calEvents');
 //const mongoose = require('mongoose');
 const { checkIdIsValid, checkString, trimName, addOnlyValidFields } = require('../utils/validate');
 
@@ -146,7 +147,7 @@ Router.delete('/:id', (req, res, next)=>{
         return next(err);
     }
     const userId = req.user.id;
-    return Patient.findOneAndRemove({"_id": id, "userId": userId})
+    return Promise.all([Patient.findOneAndRemove({"_id": id, "userId": userId}), CalEvent.deleteMany({"userId": userId, "patientId": id})])
         .then(()=>{
             return Patient.find({"userId": userId})
         })
